normalization module renamed to normalize
